--- otx/algorithms/common/adapters/mmcv/config_utils.py
@@ -1,34 +1,0 @@
-# Copyright (C) 2022 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions
-# and limitations under the License.
-
-
-from typing import Union
-
-from mmcv import Config, ConfigDict
-
-from otx.api.utils.argument_checks import (
-    check_input_parameters_type,
-)
-
-
-@check_input_parameters_type()
-def remove_from_config(config: Union[Config, ConfigDict], key: str):
-    """Update & Remove configs."""
-    if key in config:
-        if isinstance(config, Config):
-            del config._cfg_dict[key]  # pylint: disable=protected-access
-        elif isinstance(config, ConfigDict):
-            del config[key]
-        else:
-            raise ValueError(f"Unknown config type {type(config)}")
